+ Resolved compile warnings. + Added dependencies to VC80 projects - external libs/includes no longer needed. If you use other than 8th release, modify your projects accordingly.

+++ trunk/trunk_4574/src/ascent-world/HonorHandler.cpp
@@ -219,7 +219,7 @@ void HonorHandler::OnPlayerKilledUnit( Player *pPlayer, Unit* pVictim )
 				if(pAffectedPlayer->m_bg)
 					pAffectedPlayer->m_bg->HookOnHK(pAffectedPlayer);
 
-				int32 contributorpts = Points / contributors.size();
+				int32 contributorpts = int32(Points / contributors.size());
 				AddHonorPointsToPlayer(pAffectedPlayer, contributorpts);
 				if(pVictim->IsPlayer())
 				{
